big update

+++ database/migrations/2024_11_05_101049_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('price', 2);
+            $table->unsignedBigInteger('price');
             $table->text('description');
             $table->integer('stock');
             $table->string('type');
